Option to merge config files.
When the '-merge' command line option is present more than one config
file may be passed and those config files can be merged.
The test suite schema as been modified so that a param, paramGroup, testCase,
testCaseGroup, driver, driverGroup elements are root elements in addition
to testSuite. Thus a config file may exist of such an element (just like
in the case when such an element is included using the xi:include statement).

Japex will merge elements together into one testSuite. This allows one to
use a config file containing just parameters drivers and pass in the tests
cases as an addition config file, thereby ensuring that there does not need
to be a config file for each set of test cases. This can be useful when
managing large benchmarks and also for quick performance runs that may
use a subset of the tests cases.