JapexClassLoader patch
JapexClassLoader now uses the system class loader as parent as opposed
to bootstrap class loader in order for all platform classes to be
visible at runtime (such as classes from java.sql module) when running
with Java 9 or later. This is a workaround until japex is migrated to
Java 9 or later where ClassLoader.getPlatformClassLoader() can be used
as parent to provide all platform classes.